feat(auctions): move to block based pricing for auctions
Every block will have a unique price based on the provided auction settings. To support compatibility with the portal, the read function for auctions will continue to return interval based pricing. We will add a field to the response that contains the function a client can use to calculate the price of an auction at a given block height so they can generate prices locally without having the hit the service for a price for each block