feat(observations): update observations and type to store gateway and observation rollup values
Instead of storing an array of all the epochs a gateway or observer have participated in, we store two numbers (the # of time passed, and the # of times prescribed/observed) for an epoch. We additional add a third number to track the number of consecutive failures for a gateway, and this can be used to remove gateways that have failed multiple epochs in a row.