feat(read): adjust APIs for gateways
We remove the ability to set the height for prescribed observers, if someone wants to see the observers for a given height, they can leverage warp. We are not able to contruct state at given heights in the contract and the weights will be inacurrate as epochs pass the current height.

feat(read): adjust APIs for gateways

We remove the ability to set the height for prescribed observers, if someone wants to see the observers for a given height, they can leverage warp. We are not able to contruct state at given heights in the contract and the weights will be inacurrate as epochs pass the current height.