fix(epochs): add epochPeriod and update types
This also includes updating `epochStartHeight` and `epochEndHeight` in the state before distributions happen, this will make it clear (without a READ) that the current epoch value. `epochDistributionHeight` has been modified to `nextDistributionHeight` so it is also clear that while the epoch start/end may change before then, the `nextDistributionHeight` will always reflect when the next distribution will happen.